requirements.txt
colorama==0.4.4
einops==0.3.2
matplotlib==3.4.3
numpy==1.19.5
tensorflow_datasets==4.4.0
tensorflow_gpu==2.6.0
timm==0.4.12
torch==1.10.0
tqdm==4.62.3
